backend/requirements.txt
fastapi
uvicorn[standard]
pywebview
peewee
pydantic-settings
python-dotenv
anthropic
google-generativeai
httpx